Data type change for heading and

+++ dotnet/corecomponents/net.atos.daf.ct2.fms/entity/VehicleStatusResponse.cs
@@ -28,11 +28,11 @@ namespace net.atos.daf.ct2.fms.entity
     public class VehiclePositionForStatus
     {
         public int Altitude { get; set; }
-        public int Heading { get; set; }
+        public double Heading { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
         public long GPSTimestamp { get; set; }
-        public int Speed { get; set; }
+        public decimal Speed { get; set; }
     }
 }
